:art: Moved bcrypt methods in a plugin
Export will be removed, they stay for now to keep tests working
