+++ api/src/plugins/_index.ts
@@ -1,15 +1,16 @@
-import envCheck from './envCheck';
-import pgClient from './pgClient';
-import jwt from './fastifyJwt';
-import swagger from './swagger';
+import bcryptPlugin from './bcrypt';
 import cookie from './fastifyCookie';
 import cors from './fastifyCors';
+import envCheck from './envCheck';
+import jwt from './fastifyJwt';
+import pgClient from './pgClient';
+import swagger from './swagger';
 
 /**
  * **Plugins _index**
  * @description
  * Plugins are functions that are binded to the Fastify instance.
- * Add any new plugin to this array.
+ * Add any new plugin to this array (order matters!).
  */
 export const plugins = [
   envCheck, // Should always be loaded first
@@ -18,4 +19,5 @@ export const plugins = [
   cors,
   jwt,
   swagger,
+  bcryptPlugin,
 ];
